Cambio de tipos de datos en «secuenciales»
Los enteros de 16 bits ahora son enteros de 64 bits.

+++ secuenciales/src/1-5-7.rs
@@ -1,9 +1,9 @@
 use functions::read_input;
 
 fn main() {
-    let x = read_input::<i16>("Introduzca un valor para X");
-    let y = read_input::<i16>("Introduzca un valor para Y");
-    let z = read_input::<i16>("Introduzca un valor para Z");
+    let x = read_input::<i64>("Introduzca un valor para X");
+    let y = read_input::<i64>("Introduzca un valor para Y");
+    let z = read_input::<i64>("Introduzca un valor para Z");
 
     let med_geo = (x * y * z) / 3;
